Changes for Buzz-O-Matic Cheapo Edition
If you #define BOZ_ORIGINAL in boz_hw.h, you get the original pin
assignments and LCD code. If you don't, you get the pin assignments
and LCD code for the new, smaller, "Cheapo" edition of the control
unit.

The Cheapo edition doesn't have a shift register or a separate logic
gate chip, because the circuit is simplified. The buttons connect the
individual I/O pins to the interrupt pin D2. When we sleep, we make
D2 a pulled-up input and the other I/O pins output low, and if a
button is pressed that brings the interrupt line low. Then when we
wake, we reverse the arrangement - D2 becomes an output low, we make
the other pins pulled-up inputs and we can then sense which button or
buttons are pressed.

Additionally, the LCD is driven by I2C, which requires fewer I/O
lines.

There are also changes to debouncing, in both Original and Cheapo
versions of the code. The debounce timer is counted in microseconds
not milliseconds, so that people's buzz times aren't "rounded" to
the next millisecond boundary.