Attempt at better filtering out of double-steps from rotary encoder
If we sense that the rotary encoder was turned one step, less than
a certain number of milliseconds (currently 15) since the last time
the rotary encoder was turned the same direction, we don't deliver
the event to the application. This ought to help filter out those
cases where you turn the knob one step and you get two steps.