Fix irregular seconds update on chess clock app
The alarm to tell us to update the display comes whenever the running
clock's value is a multiple of 200ms. The variable amount of time it
takes to wake up from sleep and process the interrupt means that
sometimes we'll update the display when the clock is at e.g. 123.000,
and sometimes we'll update it when it's at about 122.999. This can
cause the seconds of the clock to update on the display after 800ms
or 1200ms sometimes.

Change the time we get the alarm to update the clock on the display so
we get it when the clock's value is 10ms more than a multiple of 200ms.