Slight change to last commit, and use less code
We don't want to set an alarm for e.g. 123.010 seconds if by some
unexpected delay the clock time has counted down to, say, 123.005.

General changes to reduce code size.

Finally, when we sample an unconnected analogue pin to provide a
number for a random seed, sample the analogue pin that's actually
unconnected rather than the one which senses the battery voltage.