Conundrum: make red and green buzzers be C1, and yellow and blue C2
Currently, in the conundrum app, C1 is the first buzzer (red), and C2
is all the other buzzers (green, yellow and blue). The idea behind this
is it works if you don't know whether the user will plug the buzzers
into red and green (the first two) or red and blue (the two outer
sockets).

However, if we've got two control units, and only four buzzer units
(red, green, yellow, blue) we're more likely to have red and blue on
one unit and green and yellow on the other, so green and yellow should
default to being on different sides.

As of this change, the default setting is that the first two buzzer
sockets (red and green) are player 1, and the second two buzzer sockets
(yellow and blue) are player 2.