Tweak the default buzzer noise
The Square Bell buzzer noise sounds louder for 4UP than 1UP. This is
because they use different frequencies, and the speaker seems to be
better at making the 4UP notes louder than the 1UP notes.

So I've tweaked the Square Bell buzzer noise so it sounds roughly the
same loudness for each buzzer.