Massive sweeping change for various reasons, mainly malloc/free
Dynamic memory management is now a thing. Previously, each app had to
have statically-allocated storage for all the things it could ever do,
and that all had to be in memory at once, even though not all the apps
actually run at once.

boz_mm_alloc() and boz_mm_free() therefore provide basic dynamic
allocation from a pool of 512 bytes. Any memory allocated by an
application with boz_mm_alloc() is automatically freed when that
application exits. This is accomplished by having a stack of
used-chunk-lists in the memory manager - when an app calls another app,
the used-chunk-list for the calling app is pushed onto a stack and a new
empty one is created for the called app. When the called app exits, all
the chunks in the called app's list are freed and the stack is popped,
leaving the called app with its used-chunk-list.

Note that this automatic memory-freeing only happens when the app exits,
not when it merely calls another app. The called app may therefore not
have the full 512 bytes at its disposal, because the calling app is
still there, albeit suspended, with all its dynamically allocated
storage still there.

There's also boz_mm_main_alloc(), which is only used by boz.ino's main
loop and the library functions, never by applications. This allocates
memory in a separate, special used-chunk-list which isn't affected by
apps starting and exiting.

The apps have been rewritten to use dynamically-allocated memory from
boz_mm_alloc() rather than statically-allocated storage where possible.

In addition, the struct boz_clock objects (now referred to by the
typedef pointer type boz_clock) are allocated using the dynamic memory
manager, but the main loop still keeps a static array of pointers to
them. Currently the limit on the number of clocks that can be allocated
at once is still four.

There's a "crash" app (not visible from the main menu). If you call it
from another app, with boz_crash(), the LEDs are set according to the
parameter you pass it, a sound plays, and the display shows "Guru
meditation". Nothing else is possible after that - you have to
power-cycle the Bozzard. This app is called whenever the memory manager
fails one of its many assertions.

There's a "backlight on/off" app, visible from the main menu. It toggles
the state of the display backlight and immediately exits, dropping you
back at the same point in the main menu.

The "main menu" app, which is the app that runs when you switch the
Bozzard on, now has a very basic buzzer functionality, so that you can
play with the buzzers immediately without having to select an app from
the list. If you press a buzzer, you get a buzzer noise and the other
buzzers are locked out for three seconds or until the yellow button is
pressed. The corresponding LED comes on to indicate which buzzer buzzed.

The "options" app, which is called with a struct option_menu_context *
as its argument, has a small change in addition to the large change to
make it use dynamic memory allocation. The struct option_menu_context
contains a new field, page_disable_mask, which allows the caller to
specify which entries, if any, in const struct option_page *options
should be skipped and not shown to the user.

If the bit representing (1 << page) is set in page_disable_mask, that
page isn't shown when scrolling through the option pages, and its value
in the result array is not changed.

This is useful for the conundrum app, which is also the "buzzer round"
app rebadged. If the two_sides flag is set in the rules struct (which it
is in the conundrum app but not the buzzer round app), we need to show
the "which buzzers" question in the options menu. If pressing a buzzer
stops the clock (which it does in the conundrum app but not in the
buzzer round app) then we DON'T need to show the "how many milliseconds
until we unlock the buzzers" question. Showing that, especially with an
"until yellow btn" option, in the conundrum app, is just confusing.