Vary order in which we check buzzers for events
Currently, the main loop checks all the buttons in a predictable order,
starting with buzzer 0, then buzzers 1, 2, and 3, then the operator
buttons. If we find that a button that wasn't pressed is now pressed,
we call the application's event handler for that button, if such an
event handler exists.

In the unlikely event of (say) buzzers 0 and 3 (red and blue) buzzing
at the same time, or within a few microseconds of each other so that by
the time the microcontroller has woken up and checked the I/O pins
they're both acwtive and we can't tell which was first, the red buzzer
(0) will be favoured every time simply because we check that one first.

So, let's vary the order in which we check the buttons. Each time the
main loop runs, we'll alternate between going through the button list
forwards and backwards, and after each forwards-backwards cycle we'll
change which buzzer we check first. So it goes: forwards starting with
0, backwards starting with 0, forwards starting with 1, backwards
starting with 1, and so on until we check backwards starting with 3,
after which the next check is forwards starting with 0 again.

Also: store the press and release threshold times (the durations a
button must be pressed or released before we count it as a press or a
release rather than just noise) as microseconds rather than
milliseconds. It uses a few bytes more RAM but saves dividing by 1000
which helps with keeping the code size down.

Also note: the effect where we can ramp the sound from a low frequency
to a high frequency, and we do that logarithmically, is all well and
good but that one call to pow() brings in enough math library and
floating point code that it bloats the code size by 2KB. If we run out
of code space, this is going to be first out of the window (undefine
BOZ_SOUND_RAMP_LOG to ramp the frequency linearly rather than
logarithmically).