Split sysinfo app into battery level and sysinfo, add version number
The "sysinfo" app, which used to show the date of compilation and the
battery voltage, now shows just "About"-style information: the word
Bozzard, the version number (new), a copyright line, and, if you turn
the rotary knob, the compilation date.

The new "battery" app shows the battery voltage.

There is now an additional API call, boz_get_version, which returns a
four-byte long containing the major, minor and release components of the
version number respectively in the top three bytes. The version number
is currently 0x01000000, or 1.0.0.

As of now, the compiled Bozzard code has eaten literally all of the
pies: it uses all 30,720 bytes of the available storage space on an
Arduino Nano.