Add more options to buzzer app
Make more of the rules struct accessible and tweakable by the options
menu. For the general buzzer app, the user can now set the direction of
the clock, whether the buzzes are in two teams or four players against
each other, and whether a buzz event stops the clock.

Re-order some of the elements of the rules struct and use bit-fields for
space efficiency.

Add a new, fifth buzzer noise option, "all different", which gives a
different buzzer noise to each of the four buzzers.

The logarithmic scale for the frequency-ramping thing has now been
sacrificed on the altar of code size. BOZ_SOUND_RAMP_LOG now un-defined.