Check rotary encoder on rising clock rather than falling clock
It's very easy to make the rotary encoder register a false clockwise
turn by turning it far too short a distance clockwise. We get the event
far before the turny thing reaches the next notch.

However, if we check the rotary encoder when its clock rises, rather
than when it falls, and invert the meaning of the data pin, the "edge"
seems to be far closer to the middle of the rotary encoder's travel
from notch to notch. This reduces the chance of reading a false
clockwise turn when a passing fly farts on the rotary encoder.