Persist user configuration settings to EEPROM
The Arduino Nano has some non-volatile memory in the form of an EEPROM
which the program can write to and read from. There is 1kB (0x400) of
EEPROM.

Each app has in its boz_app struct (see boz_app_list.ino) a reserved
region of the EEPROM, expressed as an offset and a length.

So far, only two apps have a reserved region of EEPROM - the Conundrum
app (64 bytes from 0x040) and the Buzzer Game app (64 bytes from 0x080).

The new functions boz_eeprom_write() and boz_eeprom_read() may be used
by an application to write and read data to and from their own section
of EEPROM only. There are also calls to get the size of the current
app's EEPROM region, and to reset the whole 1kB of EEPROM.

When the Bozzard starts up, it checks the start of EEPROM for a special
string ("BOZZARD") and a version number. If the string is there and the
version number matches the version number of the software, it carries
on. If anything's different, it calls boz_eeprom_global_reset(), which
does a "factory reset" of the whole EEPROM - it fills the whole thing
with 0xff, and writes the expected string and version number to the
start.

The conundrum and buzzer game apps use the new calls to store their
"struct game_rules" settings in the EEPROM. When a user changes a
setting in the menu and accepts it, the resulting rules struct is saved
to the EEPROM. When the app starts, if there is a valid rules struct in
the app's EEPROM region, the app uses that as the initial rules,
otherwise it uses the default rules for the app.

In all cases, EEPROM writes are performed using EEPROM.update(), which
only actually writes the byte to the EEPROM if the value currently at
that location is different. Writing a byte to EEPROM takes about 3.3ms,
and the EEPROM has a limited number of erase/write cycles, so it's
advantageous not to write a byte to EEPROM if we don't have to.