Show approximate battery level in Battery Info app
Use the voltage we sense on the battery to work out an approximate
percentage, and show a picture of a battery to depict the relevant
amount of charge.

Anything 9.3V or above is 100%, anything 7.0V or lower is 0% (because
the Arduino can't run with less than 7V on its VIN pin). For the
intermediate values I've guessed the approximate voltage at 20%, 40%,
60% and 80% based on a Duracell datasheet showing how the voltage drops
over time.

If you #define the symbol BATTERY_PICTURE to BATTERY_PICTURE_BIG, you
get a long battery symbol over five character positions. If you define
it to BATTERY_PICTURE_SMALL, you get an upright battery symbol in one
character cell. If you define it to BATTERY_PICTURE_NONE, you just get a
percentage, which while precise, may be inaccurate.

In all cases, you can press the yellow button to say you want to see the
specific percentage rather than the picture.

BATTERY_PICTURE_BIG uses the most code size, and BATTERY_PICTURE_NONE
uses the least. Currently we use BATTERY_PICTURE_BIG.

The pictures have six states: empty, 20%, 40%, 60%, 80% and full. We set
the picture to the one which is nearest the calculated percentage
battery level.