Only save buzzer options to EEPROM if "save default" is selected
Currently, every time you change an option from the buzzer game's option
menu and accept the changes, it writes those changes to EEPROM so
they're the options used next time. This might not necessarily be what
the user is expecting - they might want to change something just
temporarily but not affect the saved default settings.

So add a new option to the list of options shown by the buzzer game app,
a yes/no question "Save as default?" which is initially set to no. If,
and only if, the user changes this to yes, and accepts the changes,
write the new settings to EEPROM so they become the app's default
settings.

If the user accepts the changes without setting "Save as default" to
yes, then we apply the new rules but we don't save them to EEPROM.